==> notebooks/03_Pre-Procesado_Crear json con rangos del habla según transcripciones.ipynb <==
# Imports
import os
import json
import numpy as np
import pandas as pd
import stable_whisper

# Nuestro path de inicio siempre debe ser el home dir
os.chdir(os.path.abspath(".."))

# Archivos importantes y paths
df_annotations = pd.read_excel('data/annotations.xlsx')
dir_trans = 'data/TRANSCRIPCIONES'

# Obtengo todos los audios diferents y defino un diccionario vacío
audios_name = df_annotations['Audio_Name'].unique()
trans_dict = {}

# Loop para optener todas las tuplas de inicio fin de todas las transcripciones y guardarlas en el diccionario trans_dict
for audio_name in audios_name:
    audio_name_json = audio_name + '.json'
    audio_json = stable_whisper.WhisperResult(f'{dir_trans}/{audio_name_json}')

    x = []
    for segment in audio_json:
        x.append((segment.start, segment.end))

    trans_dict[audio_name] = x

# Verificado estan los 244 audios
print(len(trans_dict.keys()))
print(trans_dict.keys())

# Guardo el diccionario en archivo json
with open('data/intervalos_transcripciones.json', 'w') as f:
    json.dump(trans_dict, f)

Estadísticas sobre los segmentos

x = []
segmentos = []

for key in trans_dict:
    
    segmentos.append(len(trans_dict[key]))
    
    for segmento in trans_dict[key]:
        fin = segmento[1]
        inicio = segmento[0]
        
        x.append(fin - inicio)
        
print('Segmentos por audio:', np.mean(segmentos))
print('Duración promedio de segmento:', np.mean(x))
