==> notebooks/04_Pre-Procesado_Votación promedio por rango.ipynb <==
# Nuestro path de inicio siempre debe ser el home dir, e
import os
os.chdir(os.path.abspath(".."))

# Imports
import json
import pandas as pd
import stable_whisper
import warnings
from src.dataload import audio_select_mean_vote

# ignoramos warnings
warnings.filterwarnings('ignore')

# Archivos importantes y paths
df_annotations = pd.read_excel('data/annotations.xlsx')

# Omitimos todos los audios en development
df_annotations = df_annotations[df_annotations['Type'] != 'Development'].reset_index(drop = True)

# Obtenemos el json con los intervalos
with open('data/intervalos_transcripciones.json', 'r') as f:
    intervalos_transcripciones = json.load(f)
    
# Comprabamos seguimos teniendo los 244 audios
len(intervalos_transcripciones.keys())

# Audios que no estan en development
lista_audios = df_annotations['Audio_Name'].unique()
len(lista_audios)

# Creo mi diccionario que voy a guardare como json
targets_mean_vote = {}

# Loop por cada audio en la lista de audios que no estan en development
for audio_name in lista_audios:
    
    print('Procesando el audio:', audio_name)
    
    df_audio = df_annotations[df_annotations['Audio_Name'] == audio_name]
    partes = df_audio['Part_Num'].max()
    
    # Inicializo valores para el audio en el diccionario
    targets_mean_vote[audio_name] = {}
    targets_mean_vote[audio_name]['rangos'] = []
    targets_mean_vote[audio_name]['targets'] = []
    
    # Loop por cada parte del audio
    for part_num in range(partes):
        part_num = part_num + 1
        
        print(f'Procesando parte {part_num} de {audio_name}')
        
        # Obtengo votación para la parte del audio
        df_votacion = audio_select_mean_vote(df_annotations, audio_name = audio_name, part_num = part_num)
        
        # Verificamos que solo halla un start time y nos lo quedamos
        assert len(df_audio[df_audio['Part_Num'] == part_num]['start_time'].unique()) == 1, 'Más de un start time'
        start_time = df_audio[df_audio['Part_Num'] == part_num]['start_time'].unique()[0]
        
        # Agregamos al audio el start time necesario
        df_votacion['Time'] = df_votacion['Time'] + start_time
        
        # Loop principal para cargar las votaciones promedio
        i = 0
        
        int_act = intervalos_transcripciones[audio_name] 
        ts_primera_anotacion = df_votacion['Time'][0]
        ts_ultima_anotacion = df_votacion['Time'][len(df_votacion) - 1]
        
        while i + 1 <= len(int_act) and int_act[i][1] < ts_ultima_anotacion:
            
            if int_act[i][0] > ts_primera_anotacion:
                inicio = intervalos_transcripciones[audio_name][i][0]
                fin = intervalos_transcripciones[audio_name][i][1]
                
                votacion_promedio = df_votacion[(df_votacion['Time'] >= inicio) & (df_votacion['Time'] <= fin)][['Valence','Arousal','Dominance']].mean().values.tolist()
                
                targets_mean_vote[audio_name]['targets'].append(votacion_promedio)
                targets_mean_vote[audio_name]['rangos'].append((inicio, fin))
            
            i += 1
            
    assert len(targets_mean_vote[audio_name]['targets']) == len(targets_mean_vote[audio_name]['rangos']), 'Rangos no coinciden'
        
    print(f'{audio_name} procesado con éxito')
    print('')
    print('---------------------------------------------------------')

# Guardo el diccionario en archivo json
with open('data/targets_mean_vote.json', 'w') as f:
    json.dump(targets_mean_vote, f)
