==> notebooks/06_Pre-Procesado_Feature extraction (Opensmile).ipynb <==
# Nuestro path de inicio siempre debe ser el home dir, e
import os
os.chdir(os.path.abspath(".."))

# Imports
import json
import pandas as pd
import stable_whisper
import warnings
from src.dataload import load_audio_data

# ignoramos warnings
warnings.filterwarnings('ignore')

# Importación y definición del feature set a utilizar de opensmile
import opensmile

smile = opensmile.Smile(
    feature_set=opensmile.FeatureSet.ComParE_2016,
    feature_level=opensmile.FeatureLevel.Functionals,
)

opensmile_features = pd.DataFrame()

for audio_name in os.listdir('data/FEATURES/RAW/'):
    
    print(f'Procesando {audio_name}')
    
    df = pd.read_csv(f'data/FEATURES/RAW/{audio_name}')
    
    x = []    
    for index, row in df.iterrows():
        data = row['Data'].replace('[','').replace(']','').split(',')
        
        if row['Data'] != '[]':
            data = [float(i) for i in data]
            audio_features_temp = smile.process_signal(data, sampling_rate = 16000).reset_index(drop = True)
            audio_features_temp['Time'] = row['Time']
            audio_features_temp['Audio_Name'] = audio_name[:-4]
            opensmile_features = pd.concat([audio_features_temp, opensmile_features], ignore_index = True)

    print(f'{audio_name} procesado con éxito')
    print('')
    print('---------------------------------------------------------') 

opensmile_features.to_csv(f'data/FEATURES/OPENSMILE/opensmile_features.csv', index = False)
